refactor: revert to original FilePanel paste implementation
- Remove centralized paste handling from useKeyboardShortcuts
- Restore original FilePanel paste implementation for image nodes only
- Use @xyflow/react 12.8.3 preventDefault: false for better compatibility
- Improve responsibility separation: FilePanel handles node-specific paste logic
- Add specific comment about preserving browser paste events for FilePanel

Benefits:
- Better performance: paste listeners only on image nodes
- Clearer responsibility: each component handles its own paste logic
- Leverages new @xyflow/react features for conflict resolution
